[refactor] 04/06 회원 탈퇴 react query 적용 #427
