+++ app/(main)/user/components/ProfileView.tsx
@@ -10,14 +10,16 @@ import { UserDeletionModal } from "./UserDeletionModal";
 import { useRouter } from "next/navigation";
 import ProfileNickname from "./ProfileNickname";
 import ProfileImage from "./ProfileImage";
+import useDeleteUser from "./hooks/useDeleteUser";
 
 const ProfileView = () => {
-  const { user, setUser } = useAuthStore();
+  const { user } = useAuthStore();
   const profileImage = user?.profileImage || "/image/profile.svg";
   const [nicknameInput, setNicknameInput] = useState<string>("");
   const [previewImage, setPreviewImage] = useState<string>(profileImage);
   const [isDeleteModalOpen, setIsDeleteModalOpen] = useState<boolean>(false);
-  const router = useRouter();
+
+  const { mutate: deleteUser } = useDeleteUser();
 
   useEffect(() => {
     if (user?.nickname) {
@@ -27,29 +29,6 @@ const ProfileView = () => {
       setPreviewImage(user.profileImage);
     }
   }, [user?.nickname, user?.profileImage]);
-
-  const handleDeleteUser = async () => {
-    try {
-      const response = await fetch("/api/user/delete", { method: "DELETE" });
-      const data = await response.json();
-
-      if (!response.ok) {
-        alert(data.error || "회원 탈퇴에 실패했습니다.");
-        return;
-      }
-
-      alert(data.message);
-      setUser(null);
-      useAuthStore.setState({ isLoggedIn: false });
-      router.replace("/");
-    } catch (error: unknown) {
-      alert(
-        error instanceof Error
-          ? error.message
-          : "서버 오류가 발생했습니다. 다시 시도해주세요."
-      );
-    }
-  };
 
   return (
     <ProfileSection>
@@ -74,7 +53,7 @@ const ProfileView = () => {
         <UserDeletionModal
           isOpen={isDeleteModalOpen}
           onClose={() => setIsDeleteModalOpen(false)}
-          onConfirm={handleDeleteUser}
+          onConfirm={deleteUser}
         />
       )}
     </ProfileSection>
